lab2/pyproject.toml
[tool.poetry]
name = "hdr-synthesis"
version = "0.1.0"
description = "HDR image synthesis from LDR"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.10"
torch = ">=2.0.0"
torchvision = ">=0.15.0"
numpy = ">=1.24.0"
Pillow = ">=10.0.0"
opencv-python-headless = ">=4.8.0"
matplotlib = ">=3.7.0"
lpips = ">=0.1.4"
scikit-image = ">=0.21.0"
tqdm = ">=4.65.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"